Implement plugin configuration schema
In serverless v2.7.0 schema validation of the serverless.yml has been
implemented (see https://github.com/serverless/serverless/issues/8422#issuecomment-716184282)

Based on this you can trigger errors when parameters are referenced which
do not exist. This plugin relies on some custom properties in the serverless.yml
which caused warnings.

This commit appends the JSON schema as described in the README.md of this
plugin